+++ apigateway.go
@@ -8,7 +8,7 @@ import (
 	"strconv"
 	"strings"
 
-	gocf "github.com/mweagle/go-cloudformation"
+	gocf "github.com/crewjam/go-cloudformation"
 
 	"github.com/Sirupsen/logrus"
 )