+++ s3site.go
@@ -5,7 +5,7 @@ import (
 	"os"
 	"path/filepath"
 
-	gocf "github.com/mweagle/go-cloudformation"
+	gocf "github.com/crewjam/go-cloudformation"
 
 	"github.com/Sirupsen/logrus"
 	"github.com/aws/aws-sdk-go/aws"